DOMA-1803 refactor BaseMeterModalForm component
