+++ apps/condo/domains/meter/constants/constants.js
@@ -8,6 +8,8 @@ const ELECTRICITY_METER_RESOURCE_ID = '139a0d98-0631-11ec-9a03-0242ac130003'
 const HEAT_SUPPLY_METER_RESOURCE_ID = '18555734-0631-11ec-9a03-0242ac130003'
 const GAS_SUPPLY_METER_RESOURCE_ID = '1c267e92-0631-11ec-9a03-0242ac130003'
 
+const METER_MODAL_FORM_ITEM_SPAN = 11
+
 module.exports = {
     COLD_WATER_METER_RESOURCE_ID,
     HOT_WATER_METER_RESOURCE_ID,
@@ -16,5 +18,6 @@ module.exports = {
     GAS_SUPPLY_METER_RESOURCE_ID,
     CALL_METER_READING_SOURCE_ID,
     MOBILE_APP_SOURCE_ID,
+    METER_MODAL_FORM_ITEM_SPAN,
     IMPORT_CONDO_METER_READING_SOURCE_ID,
 }
