+++ apps/condo/lang/en.json
@@ -563,6 +563,9 @@
   "meterResource.GasSupply.measure": "cubic meter",
   "contact.AddContact": "Add contact",
   "meter.AddMeterReadings": "Add meter readings",
+  "meter.MeterReadingsLog": "Meter Readings",
+  "meter.modal.ChooseAccountNumber": "Choose account number",
+  "meter.modal.CreateAccountNumber": "Create new account number",
   "meter.import.MeterReading.objectsName.many": "Meter readings",
   "meter.import.MeterReading.objectsName.many.genitive": " meter readings",
   "meter.import.column.address": "Address",
@@ -588,7 +591,6 @@
   "meter.import.value.meterResourceType.electricity": "EL",
   "meter.import.value.meterResourceType.heatSupply": "HEAT",
   "meter.import.value.meterResourceType.gasSupply": "GAS",
-  "meter.MeterReadingsLog": "ИПУ",
   "Paid": "Paid",
   "Password": "Password",
   "Phone": "Phone",
